Addresses KBASE-535: "They way I read the tutorial, it suggested that the kbws-workspace command actually created the new workspace. I finally realized that wasn't happening, which is why the operation failed. Once I created the workspace in the Workspace Browser, all went well. But I think an extra step in the tutorial would be helpful."
Also:
All the hyperlinks to the command-line documentation are currently broken (see KBASE-427). One problem was that the URL of the documentation changed--that I was able to fix. However, the name anchors in cdm-entity-relationship-command-scripts are also broken and I'm not sure how to fix those.
Rather than leave dead hyperlinks, I've changed the hyperlinks to the command-line documentation to be name rather than href, so they won't do anything visible to the user but are still here in this tutorial in case the anchors ever get fixed.