Fixed link, typos, formatting.
Started out fixing the download link, then saw some typos and fixed those, noticed the page was
somewhat messy, while cleaning that up noticed some malformed tags, so fixed those, and while at
it modified the numbering for the instructions, which was a mix of ordered and unordered lists
which were also malformed. The page was run through BBEdit's "Strict Hierarchical" formatter.