Fixed bad link, repaired formatting isssues, and removed extraneous markup.
Started with a bad link on exchangeable annotations based on exemplars, as well as a styling mistake
which ensured the link would not be seen. When attempting to repair formatting problems with the
code samples, I removed the api links (which are not implemented in the api does and thus don't
work), wrapped the code in <pre><code> (rather than just <pre>), replaced <br> tags with real line
breaks (so the code samples are much easier to edit), removed the named anchors, which are not used
anywhere and just make the file hard to edit. Reformatted and checked syntax with BBEdit, verified
on the usual suspect browsers (Mac and Win).