Updating auth release notes
There are two auth repos, and two release notes. The current release
notes are labeled “auth_service_release_notes.txt”. It is confusing
because the docs directory is “auth”, but that is the name of the auth
client repo, yet the release notes are from the auth_service repo. I’m
copying into the auth docs the release notes from the auth and
auth_service repos and naming them RELEASE_NOTES-auth.txt and
RELEASE_NOTES-auth_service.txt, with the suffix being the repo.