Update kb_seed (cdmi_api) api docs, fix-perl-doc
Put the original, freshly generated from current release tag
20140224-prod; placing the source files in doc, since that is where
they are generated in the source module (kb_seed), this makes it easier
to match module source to doc source. Regenerated via fix-perl-doc
which produces a better separation of body and nav. The changes to
fix-perl-doc are better handling of options (still more to do.)