Updated docs->wordpress deploy for microbial tutorial images; removed extraneous files from same tutorial.
It is a lot easier to just deploy an entire directory of images for a tutorial. For one, the deploy script
doesn't need updating when new images are added. However, it does mean that the images directory needs to be
clean. This one had several extra js, css, image files where are not refrenced in the tutorial, they are
remnants from the old web site.
the entire directory of images.